refactor: DriverRequest DTO 통합, carColorHexCode 제거, 평점 Driver로 이전
- DriverRegisterRequest + DriverUpdateRequest → DriverRequest 통합
- DriverResponse에서 carColorHexCode 제거, averageRating 추가
- Driver 엔티티에 totalRatingSum/reviewCount 필드 및 addRating/getAverageRating 추가
- Member 엔티티에서 평점 관련 필드/메서드 제거
- ReviewCreateService, ReviewQueryService: MemberRepository → DriverRepository로 평점 조회 이전
- ProfileResponse에서 averageRating 제거
